Add image dataset handler, fix regular tester, ShowImage hook bugs. The tester is now working!

+++ wksuper/nms/__init__.py
@@ -1,6 +1,7 @@
+import numpy as np
 from .cpu_nms import cpu_nms
 
 def nms(dets, thresh):
     if dets.shape[0] == 0:
         return []
-    return cpu_nms(dets, thresh)
+    return cpu_nms(dets.astype(np.float32), thresh)
